+++ src/lib/superStruct.ts
@@ -1,14 +1,16 @@
+import type { AllKeys, MergeUnion } from './utils.js';
+
 export type SuperStructArray<T extends Record<string, unknown>, Data, ArrayData = unknown> = {
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	[Property in keyof T]?: T extends any
+	[Property in AllKeys<T>]?: [T] extends [any]
 		? NonNullable<T[Property]> extends Record<string, unknown>
-			? SuperStructArray<NonNullable<T[Property]>, Data, ArrayData>
+			? SuperStructArray<MergeUnion<NonNullable<T[Property]>>, Data, ArrayData>
 			: NonNullable<T[Property]> extends (infer A)[]
 				? ArrayData &
 						Record<
 							number | string,
 							NonNullable<A> extends Record<string, unknown>
-								? SuperStructArray<NonNullable<A>, Data, ArrayData>
+								? SuperStructArray<MergeUnion<NonNullable<A>>, Data, ArrayData>
 								: Data
 						>
 				: Data
@@ -17,12 +19,12 @@ export type SuperStructArray<T extends Record<string, unknown>, Data, ArrayData 
 
 export type SuperStruct<T extends Record<string, unknown>, Data> = Partial<{
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
-	[Property in keyof T]: T extends any
+	[Property in AllKeys<T>]: [T] extends [any]
 		? NonNullable<T[Property]> extends Record<string, unknown>
-			? SuperStruct<NonNullable<T[Property]>, Data>
+			? SuperStruct<MergeUnion<NonNullable<T[Property]>>, Data>
 			: NonNullable<T[Property]> extends (infer A)[]
 				? NonNullable<A> extends Record<string, unknown>
-					? SuperStruct<NonNullable<A>, Data>
+					? SuperStruct<MergeUnion<NonNullable<A>>, Data>
 					: Data
 				: Data
 		: never;